RFC: optee-os: provide a "debug" PACKAGECONFIG flag
This provides a simple way to enable verbose logs to check whether OP-TEE
is indeed started or not.

Note I don't include here CFG_TEE_TA_LOG_LEVEL=3, which does not have any
apparent effect, the default value of 1 not preventing TA debug messages
to be printed.

This may be a bit too low-level for this usage, though, and CORE_LOG_LEVEL=2
would be enough for the intended usage, letting the following appear on
console log to show OP-TEE starting up:

  I/TC:
  I/TC: No non-secure external DT
  I/TC: OP-TEE version: 3.11.0-dev (gcc version 10.2.0 (GCC)) #1 Fri 16 Oct 2020 02:33:06 PM UTC aarch64
  I/TC: Primary CPU initializing
  I/TC: Primary CPU switching to normal world boot

... but then "debug" would be a bad choice for a name.

It could even make sense to have CORE_LOG_LEVEL=2 be the default, if
everyone is happy that deviating from upstream on this would be good, maybe
with a PACKAGECONFIG option of "quite" which could be set at CORE_LOG_LEVEL=1
and TA_LOG_LEVEL=0 to match today's behaviour ?

Signed-off-by: Jon Mason <jon.mason@arm.com>