optee-client: start tee-supplicant with --rpmb-cid option
OP-TEE v3.17.0 introduces the command line option --rpmb-cid to
tee-supplicant [1]. This commit adds this option to the systemd
configuration to address unreliable operation with some board where
the unique eMMC device is not always assigned the same number when the
Linux kernel enumerates MMC devices.

Note that this would also be applicable to the SysV init file:
 meta-arm/recipes-security/optee/optee-client/tee-supplicant.sh
...but I have not modified it because I have not tested that
configuration.

Link: [1] https://github.com/OP-TEE/optee_client/commit/5a69d55d6596
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>